==> 1a_Map_Generation.Rmd <==
---
title: "1a_Map_generation"
author: "Alex Dhond"
date: "8/7/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

In this map I will make a map of the Neotropics that contains all the ant and dung beetle sites in the PREDICTS database

### For clearing the workspace prior to running
```{r}
rm(list=ls())
```

### Load required packages
```{r}
library(dplyr) # data manipulation
library(rnaturalearth) # generating the base map
library(rnaturalearthdata) # generating the base map data
library(ggplot2) # for plotting the map
library(ggspatial) # for adding in small map details

## You may need to install these other packages if you dont have them, for spatial data
library(sf)
library(sp)
library(rgdal)
```

### Read in the dung beetle and ant data
```{r}
## Read in all of the ant data
NeoAnts <- read.csv("NeoAnts.csv")

## Read in all of the dung beetle data
NeoDungBeetles <- read.csv("NeoDungBeetles.csv")
```

### Find the number of unique sites and studies in the data
```{r}
## Print the number of dung beetle studies
DungBeetleStudies <- NeoDungBeetles %>%
  distinct(SS, .keep_all = T)
print(paste("The number of Dung Beetle Studies in the Neotropics is", nrow(DungBeetleStudies)))

## Print the number of dung beetle sites
DungBeetleSites <- NeoDungBeetles %>%
  distinct(SSBS, .keep_all = T)
print(paste("The number of dung beetle sites in the Neotropics is", nrow(DungBeetleSites)))

## Print number of ant studies
AntStudies <- NeoAnts %>%
  distinct(SS, .keep_all = T)
print(paste("The number of Ant studies in the Neotropics is", nrow(AntStudies)))

## Print number of ant sites
AntSites <- NeoAnts %>%
  distinct(SSBS, .keep_all = T)
print(paste("The number of ant sites in the Neotropics is", nrow(AntSites)))
```

Great, so for dung beetles there are 38 studies with 1663 sites and for ants there are 55 studies and 2354 sites. NOTE: these are the total number of studies and sites within PREDICTS - I will likely filter down my data so I have less sites and studies than this

### Create a combined dataframe that has the ant and dung beetle sites
```{r}
Combined_sites <- rbind(AntSites, DungBeetleSites)
```

### Make the map
```{r}
#generate a list of all countries
countries <- ne_countries(scale = "medium", returnclass = "sf")

#subset just the Neotropical countries and combine into one dataframe
SouthAm <- countries %>%
  filter(continent == "South America")
NorthAm <- countries %>%
  filter(continent == "North America")
NorthSouth <- rbind(NorthAm, SouthAm)

## Create the map
ggplot(data = NorthSouth)+ 
  geom_sf() + 
  geom_sf(fill = "antiquewhite1") +
  coord_sf(xlim = c(-120, -30), ylim = c(-60, 35), expand = F)+
  geom_point(data = Combined_sites, aes(x = Longitude, y = Latitude, fill = Order), size = 2.5, alpha = 1, shape = 21, colour = "black")+
  theme_linedraw()+
  annotation_scale(line_width = 0.5, style = "ticks") +
  annotation_north_arrow(height = unit(1, "cm"),
  width = unit(1, "cm"),location = "bl", which_north = "true", 
        pad_x = unit(0.2, "in"), pad_y = unit(0.25, "in"),
        style = north_arrow_fancy_orienteering)+
    theme(
      text = element_text(size=10),
      plot.title = element_blank(),
      axis.title.x = element_blank(),
      axis.title.y = element_blank(),
      panel.background = element_rect(fill = "aliceblue"),
      legend.key.size = unit(0.5, "cm"),
      legend.key.width = unit(0.2,"cm"),
      legend.title = element_blank(),
      legend.spacing.y = unit(0, "mm"), 
      panel.border = element_rect(colour = "black", fill=NA),
      aspect.ratio = 1, axis.text = element_text(colour = 1, size = 12),
      legend.background = element_blank(),
      legend.box.background = element_rect(colour = "black"))+
  labs(fill = "")+
  scale_fill_discrete(labels = c("Dung beetle site", "Ant site"))+
  theme(legend.position = c(0.82, 0.9),
          legend.direction = "vertical")

## Save the map 
ggsave("site_map.png")
```
Figure X: Map of the Neotropical realm showing the distribution of ant and dung beetle sites represented in the PREDICTS database. Ant sites are depicted in blue, N = 2354. Dung beetle sites are depicted in red, N = 1663.
